# perf: Reduce DB pressure of status messages queries