--- a/migrations/frontend/1528395895_gitserver_repos_lasterror_idx.down.sql
+++ b/migrations/frontend/1528395895_gitserver_repos_lasterror_idx.down.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY gitserver_repos_last_error_idx ON gitserver_repos(last_error) WHERE last_error IS NOT NULL;