--- a/migrations/frontend/1528395895_gitserver_repos_lasterror_idx.up.sql
+++ b/migrations/frontend/1528395895_gitserver_repos_lasterror_idx.up.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+DROP INDEX IF EXISTS gitserver_repos_last_error_idx;
+
+COMMIT;