--- a/migrations/frontend/1528395896_gitserver_repos_new_lasterror_idx.down.sql
+++ b/migrations/frontend/1528395896_gitserver_repos_new_lasterror_idx.down.sql
@@ -1,0 +1,5 @@
+BEGIN;
+
+DROP INDEX IF EXISTS gitserver_repos_last_error_idx;
+
+COMMIT;