--- a/migrations/frontend/1528395896_gitserver_repos_new_lasterror_idx.up.sql
+++ b/migrations/frontend/1528395896_gitserver_repos_new_lasterror_idx.up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY gitserver_repos_last_error_idx ON gitserver_repos(repo_id) WHERE last_error IS NOT NULL;